+++ src/parser/enrichers/monster/Generic/_module.mjs
@@ -7,3 +7,4 @@ export { default as ReversalOfFortune } from "./ReversalOfFortune.mjs";
 export { default as SuaveDefense } from "./SuaveDefense.mjs";
 export { default as NamedEffect } from "./NamedEffect.mjs";
 export { default as Illumination } from "./Illumination.mjs";
+export { default as Evasion } from "./Evasion.mjs";