+++ src/parser/enrichers/monster/HomunculusServant/_module.mjs
@@ -1,0 +1,1 @@
+export { default as MagicBond } from "./MagicBond.mjs";